Добавленны зависимости для  RxJava Поправил типы данных для модели Json Добавили Модели для GeoCodingDataModel для получения названия города по lat lon Добавили GeoCodingApi , теперь оба API по рузальтату запроса отдабт Observable<myModel>
Добавили BaseRepository.kt c обьектом dataEmitter (дата Излучатель)
Добавили MainRepository.kt c model data class который и будет отправлятся в презентер из метода reloadData Который RxJAva:
     упаковывается два запроса к серверу, получает результаты в лямбду
     запукскается на IO потоке
     сохранит  в БД
     в случае ошибки Достанет из БД
     подписываем 2 подписчика,
      1-й подп если не было ошибок
     2-й подп если ошибка

MainPresenter.kt теперь имеет ссылку на MainRepository.kt
 Fun Enable подписывает датаЕмиттер на ServerResponse? и при полученнии обновляет состояние viewState и View соответвенно

+++ app/src/main/java/ru/weather/weathertoday/busines/model/WeatherDataModel.kt
@@ -4,8 +4,8 @@ data class WeatherDataModel(
     val current: Current,
     val daily: List<DailyWeatherModel>,
     val hourly: List<HourlyWeatherModel>,
-    val lat: Int,
-    val lon: Int,
+    val lat: Double,
+    val lon: Double,
     val timezone: String,
     val timezone_offset: Int
 )
